Better Better Tower Placement System
added tower selection menu for which towers the user wants to place down.
beside the menu, a button to switch states - for upgrading towers.
Bugs:
Towers can still be placed on walls, in drill, player, enemy spawns, and other towers. Towers placed on player makes  the player clip out of the map.